Add access control and authentication (breaking backup function)

+++ src/main/java/nl/taallijn/has/http/HttpServerVerticle.java
@@ -19,13 +19,24 @@ import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
+import io.vertx.ext.auth.AuthProvider;
+import io.vertx.ext.auth.shiro.ShiroAuth;
+import io.vertx.ext.auth.shiro.ShiroAuthOptions;
+import io.vertx.ext.auth.shiro.ShiroAuthRealmType;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.client.HttpResponse;
 import io.vertx.ext.web.client.WebClient;
 import io.vertx.ext.web.client.WebClientOptions;
 import io.vertx.ext.web.codec.BodyCodec;
+import io.vertx.ext.web.handler.AuthHandler;
 import io.vertx.ext.web.handler.BodyHandler;
+import io.vertx.ext.web.handler.CookieHandler;
+import io.vertx.ext.web.handler.FormLoginHandler;
+import io.vertx.ext.web.handler.RedirectAuthHandler;
+import io.vertx.ext.web.handler.SessionHandler;
+import io.vertx.ext.web.handler.UserSessionHandler;
+import io.vertx.ext.web.sstore.LocalSessionStore;
 import io.vertx.ext.web.templ.FreeMarkerTemplateEngine;
 import nl.taallijn.has.database.WikiDatabaseService;
 
@@ -48,14 +59,35 @@ public class HttpServerVerticle extends AbstractVerticle {
 		HttpServer server = vertx.createHttpServer(new HttpServerOptions().setSsl(true)
 				.setKeyStoreOptions(new JksOptions().setPath("server-keystore.jks").setPassword("secret")));
 
+		AuthProvider auth = ShiroAuth.create(vertx, new ShiroAuthOptions().setType(ShiroAuthRealmType.PROPERTIES)
+				.setConfig(new JsonObject().put("properties_path", "classpath:wiki-users.properties")));
+
 		Router router = Router.router(vertx);
+
+		router.route().handler(CookieHandler.create());
+		router.route().handler(BodyHandler.create());
+		router.route().handler(SessionHandler.create(LocalSessionStore.create(vertx)));
+		router.route().handler(UserSessionHandler.create(auth));
+
+		AuthHandler authHandler = RedirectAuthHandler.create(auth, "/login");
+		router.route("/").handler(authHandler);
+		router.route("/wiki/*").handler(authHandler);
+		router.route("/action/*").handler(authHandler);
+
 		router.get("/").handler(this::indexHandler);
-		router.get("/backup").handler(this::backupHandler);
 		router.get("/wiki/:page").handler(this::pageRenderingHandler);
-		router.post().handler(BodyHandler.create());
-		router.post("/save").handler(this::pageUpdateHandler);
-		router.post("/create").handler(this::pageCreateHandler);
-		router.post("/delete").handler(this::pageDeletionHandler);
+		router.post("/action/save").handler(this::pageUpdateHandler);
+		router.post("/action/create").handler(this::pageCreateHandler);
+		router.post("/action/delete").handler(this::pageDeletionHandler);
+		router.get("/action/backup").handler(this::backupHandler);
+
+		router.get("/login").handler(this::loginHandler);
+		router.post("/login-auth").handler(FormLoginHandler.create(auth));
+
+		router.get("/logout").handler(context -> {
+			context.clearUser();
+			context.response().setStatusCode(302).putHeader("Location", "/").end();
+		});
 
 		Router apiRouter = Router.router(vertx);
 		apiRouter.get("/pages").handler(this::apiRoot);
@@ -187,64 +219,89 @@ public class HttpServerVerticle extends AbstractVerticle {
 		});
 	}
 
-	private void indexHandler(RoutingContext context) {
-		dbService.fetchAllPages(reply -> {
-			if (reply.succeeded()) {
-				context.put("title", "Wiki home");
-				context.put("pages", reply.result().getList());
-				templateEngine.render(context, "templates", "/index.ftl", ar -> {
-					if (ar.succeeded()) {
-						context.response().putHeader("Content-Type", "text/html");
-						context.response().end(ar.result());
-					} else {
-						context.fail(ar.cause());
-					}
-				});
+	private void loginHandler(RoutingContext context) {
+		context.put("title", "Login");
+		templateEngine.render(context, "templates", "/login.ftl", ar -> {
+			if (ar.succeeded()) {
+				context.response().putHeader("Content-Type", "text/html");
+				context.response().end(ar.result());
 			} else {
-				context.fail(reply.cause());
+				context.fail(ar.cause());
 			}
 		});
 	}
 
+	private void indexHandler(RoutingContext context) {
+		context.user().isAuthorized("create", res -> {
+			boolean canCreatePage = res.succeeded() && res.result();
+			dbService.fetchAllPages(reply -> {
+				if (reply.succeeded()) {
+					context.put("title", "Wiki home");
+					context.put("pages", reply.result().getList());
+					context.put("canCreatePage", canCreatePage);
+					context.put("username", context.user().principal().getString("username"));
+					templateEngine.render(context, "templates", "/index.ftl", ar -> {
+						if (ar.succeeded()) {
+							context.response().putHeader("Content-Type", "text/html");
+							context.response().end(ar.result());
+						} else {
+							context.fail(ar.cause());
+						}
+					});
+				} else {
+					context.fail(reply.cause());
+				}
+			});
+		});
+	}
+
 	private void backupHandler(RoutingContext context) {
-		dbService.fetchAllPagesData(reply -> {
-			if (reply.succeeded()) {
-				JsonArray filesObject = new JsonArray();
-				JsonObject payload = new JsonObject().put("files", filesObject).put("language", "plaintext")
-						.put("title", "vertx-wiki-backup").put("public", true);
-				reply.result().forEach(page -> {
-					JsonObject fileObject = new JsonObject();
-					fileObject.put("name", page.getString(actualFieldName(page, "NAME")));
-					fileObject.put("content", page.getString(actualFieldName(page, "CONTENT")));
-					filesObject.add(fileObject);
-				});
-				webClient.post(443, "snippets.glot.io", "/snippets").putHeader("Content-Type", "application/json")
-						.as(BodyCodec.jsonObject()).sendJsonObject(payload, ar -> {
-							if (ar.succeeded()) {
-								HttpResponse<JsonObject> response = ar.result();
-								if (response.statusCode() == 200) {
-									String url = "https://glot.io/snippets/" + response.body().getString("id");
-									context.put("backup_gist_url", url);
-									indexHandler(context);
-								} else {
-									StringBuilder message = new StringBuilder().append("Could not backup the wiki: ")
-											.append(response.statusMessage());
-									JsonObject body = response.body();
-									if (body != null) {
-										message.append(System.getProperty("line.separator"))
-												.append(body.encodePrettily());
-									}
-									LOGGER.error(message.toString());
-									context.fail(502);
-								}
-							} else {
-								Throwable err = ar.cause();
-								LOGGER.error("HTTP Client error", err);
-								context.fail(err);
-							}
+		context.user().isAuthorized("role:writer", res -> {
+			if (res.succeeded() && res.result()) {
+				dbService.fetchAllPagesData(reply -> {
+					if (reply.succeeded()) {
+						JsonArray filesObject = new JsonArray();
+						JsonObject payload = new JsonObject().put("files", filesObject).put("language", "plaintext")
+								.put("title", "vertx-wiki-backup").put("public", true);
+						reply.result().forEach(page -> {
+							JsonObject fileObject = new JsonObject();
+							fileObject.put("name", page.getString(actualFieldName(page, "NAME")));
+							fileObject.put("content", page.getString(actualFieldName(page, "CONTENT")));
+							filesObject.add(fileObject);
 						});
+						webClient.post(443, "snippets.glot.io", "/snippets")
+								.putHeader("Content-Type", "application/json").as(BodyCodec.jsonObject())
+								.sendJsonObject(payload, ar -> {
+									if (ar.succeeded()) {
+										HttpResponse<JsonObject> response = ar.result();
+										if (response.statusCode() == 200) {
+											String url = "https://glot.io/snippets/" + response.body().getString("id");
+											context.put("backup_gist_url", url);
+											indexHandler(context);
+										} else {
+											StringBuilder message = new StringBuilder()
+													.append("Could not backup the wiki: ")
+													.append(response.statusMessage());
+											JsonObject body = response.body();
+											if (body != null) {
+												message.append(System.getProperty("line.separator"))
+														.append(body.encodePrettily());
+											}
+											LOGGER.error(message.toString());
+											context.fail(502);
+										}
+									} else {
+										Throwable err = ar.cause();
+										LOGGER.error("HTTP Client error", err);
+										context.fail(err);
+									}
+								});
+					} else {
+						context.fail(reply.cause());
+					}
+				});
 			} else {
-				context.fail(reply.cause());
+				context.response().setStatusCode(403).end();
 			}
 		});
 	}
@@ -255,49 +312,65 @@ public class HttpServerVerticle extends AbstractVerticle {
 	}
 
 	private void pageRenderingHandler(RoutingContext context) {
-		String requestedPage = context.request().getParam("page");
-		dbService.fetchPage(requestedPage, reply -> {
-			if (reply.succeeded()) {
-				JsonObject payLoad = reply.result();
-				boolean found = payLoad.getBoolean("found");
-				String rawContent = payLoad.getString("rawContent", EMPTY_PAGE_MARKDOWN);
-				context.put("title", requestedPage);
-				context.put("id", payLoad.getInteger("id", -1));
-				context.put("newPage", found ? "no" : "yes");
-				context.put("rawContent", rawContent);
-				context.put("content", Processor.process(rawContent));
-				context.put("timestamp", new Date().toString());
-				templateEngine.render(context, "templates", "/page.ftl", ar -> {
-					if (ar.succeeded()) {
-						context.response().putHeader("Content-Type", "text/html");
-						context.response().end(ar.result());
+		context.user().isAuthorized("update", updateResponse -> {
+			boolean canSavePage = updateResponse.succeeded() && updateResponse.result();
+			context.user().isAuthorized("delete", deleteResponse -> {
+				boolean canDeletePage = deleteResponse.succeeded() && deleteResponse.result();
+				String requestedPage = context.request().getParam("page");
+				dbService.fetchPage(requestedPage, reply -> {
+					if (reply.succeeded()) {
+						JsonObject payLoad = reply.result();
+						boolean found = payLoad.getBoolean("found");
+						String rawContent = payLoad.getString("rawContent", EMPTY_PAGE_MARKDOWN);
+						context.put("title", requestedPage);
+						context.put("id", payLoad.getInteger("id", -1));
+						context.put("newPage", found ? "no" : "yes");
+						context.put("rawContent", rawContent);
+						context.put("content", Processor.process(rawContent));
+						context.put("timestamp", new Date().toString());
+						context.put("username", context.user().principal().getString("username"));
+						context.put("canSavePage", canSavePage);
+						context.put("canDeletePage", canDeletePage);
+						templateEngine.render(context, "templates", "/page.ftl", ar -> {
+							if (ar.succeeded()) {
+								context.response().putHeader("Content-Type", "text/html");
+								context.response().end(ar.result());
+							} else {
+								context.fail(ar.cause());
+							}
+						});
 					} else {
-						context.fail(ar.cause());
+						context.fail(reply.cause());
 					}
 				});
-			} else {
-				context.fail(reply.cause());
-			}
+			});
 		});
 	}
 
 	private void pageUpdateHandler(RoutingContext context) {
-		String title = context.request().getParam("title");
-		Handler<AsyncResult<Void>> handler = reply -> {
-			if (reply.succeeded()) {
-				context.response().setStatusCode(303);
-				context.response().putHeader("Location", "/wiki/" + title);
-				context.response().end();
+		boolean pageCreation = "yes".equals(context.request().getParam("newPage"));
+		context.user().isAuthorized(pageCreation ? "create" : "update", res -> {
+			if (res.succeeded() && res.result()) {
+				String title = context.request().getParam("title");
+				Handler<AsyncResult<Void>> handler = reply -> {
+					if (reply.succeeded()) {
+						context.response().setStatusCode(303);
+						context.response().putHeader("Location", "/wiki/" + title);
+						context.response().end();
+					} else {
+						context.fail(reply.cause());
+					}
+				};
+				String markdown = context.request().getParam("markdown");
+				if (pageCreation) {
+					dbService.createPage(title, markdown, handler);
+				} else {
+					dbService.savePage(Integer.valueOf(context.request().getParam("id")), markdown, handler);
+				}
 			} else {
-				context.fail(reply.cause());
+				context.response().setStatusCode(403).end();
 			}
-		};
-		String markdown = context.request().getParam("markdown");
-		if ("yes".equals(context.request().getParam("newPage"))) {
-			dbService.createPage(title, markdown, handler);
-		} else {
-			dbService.savePage(Integer.valueOf(context.request().getParam("id")), markdown, handler);
-		}
+		});
 	}
 
 	private void pageCreateHandler(RoutingContext context) {
@@ -312,13 +385,19 @@ public class HttpServerVerticle extends AbstractVerticle {
 	}
 
 	private void pageDeletionHandler(RoutingContext context) {
-		dbService.deletePage(Integer.valueOf(context.request().getParam("id")), reply -> {
-			if (reply.succeeded()) {
-				context.response().setStatusCode(303);
-				context.response().putHeader("Location", "/");
-				context.response().end();
+		context.user().isAuthorized("delete", res -> {
+			if (res.succeeded() && res.result()) {
+				dbService.deletePage(Integer.valueOf(context.request().getParam("id")), reply -> {
+					if (reply.succeeded()) {
+						context.response().setStatusCode(303);
+						context.response().putHeader("Location", "/");
+						context.response().end();
+					} else {
+						context.fail(reply.cause());
+					}
+				});
 			} else {
-				context.fail(reply.cause());
+				context.response().setStatusCode(403).end();
 			}
 		});
 	}
